Pre-implemented TradingStrategiesTester.

+++ TradingBot.Core/Domain/Chart/ChartDataItem.cs
@@ -8,9 +8,9 @@ namespace TradingBot.Core.Domain.Chart
 
         public IQuote Quote { get; }
 
-        public IDictionary<string, double?> Indicators { get; }
+        public IDictionary<string, decimal?> Indicators { get; }
 
-        public ChartDataItem(DateTime date, IQuote quote, IDictionary<string, double?> indicators)
+        public ChartDataItem(DateTime date, IQuote quote, IDictionary<string, decimal?> indicators)
         {
             if (quote == null) throw new ArgumentNullException(nameof(quote));
             if (indicators == null) throw new ArgumentNullException(nameof(indicators));
